How To Update The Sources List
We are currently working on fixing the recent issue caused by Ubuntu 22.10's end of life, so we made sure to give you instructions on how you can restore the operating system's ability to install & update packages. Meanwhile, we are trying our best to release versions of our operating systems that have the repositories made for the latest versions of Ubuntu.